backend/requirements.txt
Django==5.1.3
djangorestframework
psycopg2-binary
gunicorn
whitenoise
python-dotenv
django-cors-headers==4.3.1
Pillow>=10.1.0  # Added this line